+++ src/ethereum/networkAdapters/networkAdapterTypes.ts
